+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * Add `api::projects::DeleteProject` endpoint
   * Add `api::projects::deployments::DeleteDeployment` endpoint
+  * Add `api::projects::deployments::Deployment` endpoint
   * Add `api::projects::deployments::EditDeployment` endpoint
 
 # v0.1700.1